+++ AdaptationLayer/src/eu/qualimaster/adaptation/external/ReplayMessage.java
@@ -31,7 +31,7 @@ public class ReplayMessage extends RequestMessage {
     private int ticket;
     private Date start;
     private Date end;
-    private int speed;
+    private float speed;
     private String query;
     
     /**
@@ -60,7 +60,7 @@ public class ReplayMessage extends RequestMessage {
      * @param speed the replay speed, positive speed up, negative slow down
      * @param query which data to send (format unspecified so far)
      */
-    public void setReplayStartInfo(Date start, Date end, int speed, String query) {
+    public void setReplayStartInfo(Date start, Date end, float speed, String query) {
         this.start = start;
         this.end = end;
         this.speed = speed;
@@ -126,7 +126,7 @@ public class ReplayMessage extends RequestMessage {
      * 
      * @return the replay speed
      */
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
